refactor: to test provider
