+++ packages/react-broadcast/src/components/index.ts
@@ -1,4 +1,4 @@
 export { default as ToggleFeature } from './toggle-feature';
 export { default as ConfigureFlopFlip } from './configure';
 export { default as ReconfigureFlopFlip } from './reconfigure';
-export { default as FlopflipTestContextProvider } from './test-context-provider';
+export { default as FlopflipTestProvider } from './test-provider';
