--- packages/react-broadcast/src/components/test-context-provider/index.ts
@@ -1,1 +1,0 @@
-export { TestContextProvider as default } from './test-context-provider';
