+++ packages/react-broadcast/src/components/test-provider/index.ts
@@ -1,0 +1,1 @@
+export { TestProvider as default } from './test-provider';
